suppress warnings
git-svn-id: https://svn.salilab.org/imp/trunk@16768 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/algebra/include/Transformation3D.h
@@ -140,8 +140,8 @@ IMPALGEBRAEXPORT Transformation3D get_transformation_3d(
   **/
 IMPALGEBRAEXPORT Transformation3D get_random_local_transformation(
    Vector3D origin,
-   float max_translation=5.,
-   float max_angle_in_rad=0.26);
+   double max_translation=5.,
+   double max_angle_in_rad=0.26);
