+++ modules/algebra/src/Triangle3D.cpp
@@ -13,9 +13,9 @@ namespace {
 //! Return true if the vectors are paralel
 bool get_are_parallel(const Vector3D &v1,
                                const Vector3D &v2) {
-  float dot_p = v1*v2;
-  float det1 = v1.get_magnitude();
-  float det2 = v2.get_magnitude();
+  double dot_p = v1*v2;
+  double det1 = v1.get_magnitude();
+  double det2 = v2.get_magnitude();
   return (std::abs(std::abs(dot_p)-std::abs(det1*det2))<0.0001);
 }
 }
